application fixes
