+++ Engine/Source/Fusion/Public/Fusion.h
@@ -2,10 +2,14 @@
 
 #include "FusionCore.h"
 
+// - Windows -
 #include "Window/FToolWindow.h"
+
+// - Styles -
 #include "Style/FToolWindowStyle.h"
 #include "Style/FListViewStyle.h"
 #include "Style/FExpandableSectionStyle.h"
+#include "Style/FTreeViewStyle.h"
 
 // - Data -
 #include "Data/FAbstractItemModel.h"
@@ -19,7 +23,6 @@
 #include "Widget/ListView/FListView.h"
 
 // - Tree View -
-#include "Widget/TreeView/FTreeViewStyle.h"
 #include "Widget/TreeView/FTreeViewItem.h"
 #include "Widget/TreeView/FTreeViewContainer.h"
 #include "Widget/TreeView/FTreeViewHeaderColumn.h"
@@ -29,4 +32,6 @@
 #include "Widget/TreeView/FTreeViewModel.h"
 #include "Widget/TreeView/FTreeView.h"
 
+// - Other Widgets -
 #include "Widget/FExpandableSection.h"
+
